Vector field gravity working
Refined the scripts and got vector field gravity working for 12x12x12
matricies
Needs to be optimized very badly, perhaps compute shaders
Also mask out the internal vectors using colliders

+++ Assets/Scripts/Planetoid.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class Planetoid : MonoBehaviour 
 {
 	public Transform coordinateSystem;
-	public int mass;
+	public float mass;
 	
 	// Use this for initialization
 	void Start()
